backend/requirements.txt
langchain
langchain-community
langchain-ollama
langchain-chroma
chromadb
unstructured
networkx
pdfminer.six
pi_heif
unstructured-inference
pdf2image
pytesseract
pypdf
reportlab
httpx
python-jose[cryptography]
python-dotenv
pandas
openpyxl
docx2txt
passlib[bcrypt]